+++ src/components/DishInfo.vue
